# Conflicts:
#	Assets/_Scripts/Entities/Player.cs
#	ProjectSettings/TagManager.asset

+++ Assets/_Scripts/Entities/Enemies/ShamanEnemy.cs
@@ -3,7 +3,7 @@
     /// <summary>Checks if the entity can be attacked, and attacks them if so.</summary>
     /// <param name="target">The entity to attack.</param>
     /// <param name="damage">The damage to deal to the entity.</param>
-    public override void Attack(Entity target, int damage)
+    public override void Attack(Entity target, float damage)
     {
         base.Attack(target, damage);
     }
