+++ Assets/_Scripts/Entities/Enemy.cs
@@ -9,7 +9,7 @@ public abstract class Enemy : Entity, IFreezable
     /// <summary>Checks if the entity can be attacked, and attacks them if so.</summary>
     /// <param name="target">The entity to attack.</param>
     /// <param name="damage">The damage to deal to the entity.</param>
-    public override void Attack(Entity target, int damage)
+    public override void Attack(Entity target, float damage)
     {
         if(!(target is Player))
             return;
