+++ Assets/_Scripts/Entities/Entity.cs
@@ -9,7 +9,7 @@ public abstract class Entity : MonoBehaviour
     /// <summary>Checks if the entity can be attacked, and attacks them if so.</summary>
     /// <param name="target">The entity to attack.</param>
     /// <param name="damage">The damage to deal to the entity.</param>
-    public virtual void Attack(Entity target, int damage)
+    public virtual void Attack(Entity target, float damage)
     {
         Health targetHp = target.GetComponent<Health>();
